Struct blstrs::Fp

#[repr(transparent)]
pub struct Fp(_);

Fp values are always in Montgomery form; i.e., Fp(a) = aR mod p, with R = 2^384. blst_fp.l is in little-endian u64 limbs format.

Implementations

impl Fp

pub fn char() -> [u8; 48]

pub fn from_bytes_le(bytes: &[u8; 48]) -> CtOption<Fp>

Attempts to convert a little-endian byte representation of a scalar into an Fp, failing if the input is not canonical.

pub fn from_bytes_be(be_bytes: &[u8; 48]) -> CtOption<Fp>

Attempts to convert a big-endian byte representation of a scalar into an Fp, failing if the input is not canonical.

pub fn to_bytes_le(&self) -> [u8; 48]

Converts an element of Fp into a byte representation in little-endian byte order.

pub fn to_bytes_be(&self) -> [u8; 48]

Converts an element of Fp into a byte representation in big-endian byte order.

pub fn from_raw_unchecked(l: [u64; 6]) -> Fp

Constructs an element of Fp from a little-endian array of limbs without checking that it is canonical and without converting it to Montgomery form (i.e. without multiplying by R).

pub fn mul3(&self) -> Self

Multiplies self with 3, returning the result.

pub fn mul8(&self) -> Self

Multiplies self with 8, returning the result.

pub fn shl(&self, count: usize) -> Self

Left shift self by count, returning the result.

pub fn from_u64s_le(bytes: &[u64; 6]) -> CtOption<Self>

pub fn num_bits(&self) -> u32

pub fn is_quad_res(&self) -> Choice

pub fn square_assign(&mut self)